# Huge speedup of the classical calculator for few sites