While working at https://github.com/gem/oq-engine/issues/7745 I discovered that a *huge* amount of time was wasted in computing `len(ctx)` millions of time.
```
$ OQ_SAMPLE_SITES=.001 oq run GLD.zip -c0 -s100 && oq show performance

# master
| calc_55097, maxmem=0.4 GB  | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| ClassicalCalculator.run    | 134.1    | 128.6     | 1      |
| total classical            | 126.6    | 204.8     | 1      |
| make_contexts              | 56.4     | 187.6     | 2_516  |

# this branch, double speed
| calc_55098, maxmem=0.4 GB  | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| ClassicalCalculator.run    | 70.7     | 129.1     | 1      |
| total classical            | 63.4     | 203.8     | 1      |
| make_contexts              | 54.4     | 187.8     | 2_516  |
```
The effect is huge in the case of few sites, but visible also in the case of many sites.